Bridge adjustments. Getting rid of 4096 non-completed requests limitation.
1. Use primitive hash map implementation instead with linked-list conflict resolution.
2. Add some synchronization for request id counter, so it's guaranteed to be atomically incremented.
3. Passing custom callback as a 3rd argument of `tc_bridge_request` - still a question how to use it on the client side.

1. Still some theoretical issues with the highly concurrent environment, where more than 4096 requests are created
_at the same moment_ which is almost not possible in my understanding.

2. Another issue may be when we're exceeding the max request ID (which is `UINT32_MAX` =4'294'967'295).

`tonclient.h` should be modified so instead of passing `uint32_t request_id` we could pass `void* ptr`.

This would eliminate the need of transforming callback data structures to `uint32_t` and back, thus removing the need of hash table logic
as all is needed in this case is to pass that pointer to the `struct` containing all the callback info and then free it when request processing is finished.